Added rotational velocity (torque) Renamed _gameTime to _deltaTime Changed for loops with vector size to increment with a size_t variable instead of an int

+++ GameLibrary/FileWriter.cpp
@@ -11,7 +11,7 @@ FileWriter::~FileWriter() {
 void FileWriter::writeToFile(std::string _fileName, std::vector<std::string> _data){
 	std::ofstream file;
 	file.open(_fileName);
-	for (int i = 0; i < _data.size(); i++){
+	for (size_t i = 0; i < _data.size(); i++){
 		file << _data[i];
 	}
 	file.close();
